mystery-gift/requirements.txt
toml==0.10.2
websocket-client==1.8.0
